+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0.1
 * Add `reconnectOnClose` and `reconnectinterval` parameters in top-level `WebSockets` constructors.
 * Close `WebSocketExtraneousEventHandler`.
+* Add onAuthenticated to server-side.
 
 # 2.0.0
 * Update to work with `client@2.0.0`.